+++ src/components/grid/editComponent.js
@@ -12,9 +12,11 @@ import {
 	Content,
     Footer,
     Item,
-    Input
+    Input,
+    Title
+
 } from "native-base";
-import CustomHeader from '../headers/header'
+import CustomHeader from '../headers/hearder'
 import styles from "./styles";
 import {isEmpty} from "../../utils/validators";
 
@@ -25,39 +27,42 @@ const EditComponent = props => {
         onChangeText,
         handleOnSaveClick,
     } = props
+
     
 	renderItemList =(rowData)=>{
         return rowData ? (
-            <View style ={{flex:1}} >
-                <Text style={styles.topLabel}
-                > 
-                    {rowData[0]['rowKey']||''}
-                </Text>
+            <View style={{ borderBottomWidth: 0, paddingVertical: 5 }} >
+                <View style={{marginVertical:20,marginHorizontal:10}}>
+                    <Title style={styles.topLabel}> 
+                        {rowData[0]['rowKey']||''}
+                    </Title>
+                </View>
+                
                 {rowData.map((item, i) =>{
                     const keyboardType =  item['type'] && item['type'].toLowerCase()==='number'? 'numeric': 'default';
                     const editable = !item['editable'] ? item['editable'] : true;
                     return(
-                        <Item
-                            key ={`${item['rowKey']+""+item['colKey']}`}
-                            style={{height:40,width:'90%',marginBottom:5}}>
-                            <Input
-                                style={{
-                                    height:Platform.OS === 'ios' ? undefined : null,
-                                    paddingStart:5,
-                                }}
-                                underlineColorAndroid="transparent"
-                                numberOfLines={5}
-                                maxLength={100}
-                                multiline={true}
-                                placeholder={item['colKey']||'None'}
-                                placeholderTextColor={theme.inputColorPlaceholder}
-                                onChangeText={text => onChangeText(item['rowKey'] ,item['colKey'],text)}
-                                value={item['value']}
-                                keyboardType={keyboardType}
-                                editable={editable}
-                            
-                            />
-                        </Item>
+                        <View key ={`${item['rowKey']+""+item['colKey']}`} style={{flex:1,marginBottom:5}}>
+                            <Item style={{height:50}}>
+                                <Input
+                                    style={{
+                                        height:Platform.OS === 'ios' ? undefined : null,
+                                        paddingStart:5,
+                                    }}
+                                    underlineColorAndroid="transparent"
+                                    numberOfLines={5}
+                                    maxLength={100}
+                                    multiline={true}
+                                    placeholder={item['colKey']||'None'}
+                                    placeholderTextColor={theme.inputColorPlaceholder}
+                                    onChangeText={text => onChangeText(item['rowKey'] ,item['colKey'],text)}
+                                    value={item['value']}
+                                    keyboardType={keyboardType}
+                                    editable={editable}
+                                
+                                />
+                            </Item>
+                        </View>
                     )})
                 } 
             </View>
@@ -67,7 +72,7 @@ const EditComponent = props => {
 	return (
 		<Container style={{ flex: 1 }}>
 				<CustomHeader {...props } />
-				<Content style={{marginBottom:150}}>
+				<Content style={{marginBottom:50}}>
 					<View style={styles.container}>
 						{!isEmpty(rowData) && Array.isArray(rowData)? renderItemList(rowData): null}
 					</View>
