+++ src/components/grid/styles.js
@@ -191,18 +191,13 @@ summaryText:{
 
 },
 topLabel:{
-  alignSelf:'center',
-  fontFamily: 'roboto',
   fontSize:18,
-  fontWeight: '300',
-  color: '#989898',
-  borderBottomColor: '#d6d7da',
-  borderBottomWidth: 0.5,
+  color: 'rgb(0,151,235)',
+  alignSelf:'flex-start',
   minWidth:'18%',
-  paddingHorizontal:5,
-  textAlign:'center'
+  paddingHorizontal:15,
+  textAlign:'center',
 }
-
 });
 
 export default styles;
